GLK: FROTZ: Shift Header class into it's own file
